reorginize model.py and network.py
